Merged PR 1346744: handle unreconciled account notification
handle unreconciled account notification

DG will only support a single account per tenant which will be the reconciled account. If an account is unreconciled then it should be treated as a delete so that all processing and data is removed.

When a new primary reconciled account is selected, that one will take its place. No  active accounts for the same tenant can not exist at the same time.

Related work items: #3176660