Merged PR 1437926: handle unreconciled account notification
1. handle unreconciled account notification

DG will only support a single account per tenant which will be the reconciled account. If an account is unreconciled then it should be treated as a delete so that all processing and data is removed.

When a new primary reconciled account is selected, that one will take its place. 2 active accounts for the same tenant can not exist at the same time.

2. Enforce creating of admin settings

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix to handle unreconciled account notifications.

#### PR Summary
This pull request addresses the handling of unreconciled account notifications by adding new methods and refactoring existing logic.
- `src/ApiService/Controllers/ControlPlane/PlatformAccountNotificationsController.cs`: Added `Create`, `Delete`, and `IsInvalidReconcileState` methods; refactored existing methods to use these new methods for better handling of account states.
- `src/Core/Services/CatalogConfig/CatalogConfigService.cs`: Fixed a bug to ensure `Features` are correctly retained during updates.